requirements.txt
Flask==2.3.2
Pillow==10.0.0
requests==2.31.0